requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0
numpy>=1.24.0
tqdm>=4.65.0
pydantic>=2.0.0
plotly>=5.17.0
kaleido>=0.2.1
pytest>=7.4.0
pytest-mock>=3.11.0
pytest-cov>=4.1.0

